Revert "Nettobetrag und Steuerbetrag zu BigDecimal geändert"
This reverts commit 63d161ab55f70ca50b0ff95ad55ba8a3cecf3174.

+++ src/de/jost_net/JVerein/rmi/SollbuchungPosition.java
@@ -16,7 +16,6 @@
  **********************************************************************/
 package de.jost_net.JVerein.rmi;
 
-import java.math.BigDecimal;
 import java.rmi.RemoteException;
 import java.util.Date;
 
@@ -57,8 +56,8 @@ public interface SollbuchungPosition extends DBObject
 
   public String getZweck() throws RemoteException;
 
-  public BigDecimal getNettobetrag() throws RemoteException;
+  public Double getNettobetrag() throws RemoteException;
 
-  public BigDecimal getSteuerbetrag() throws RemoteException;
+  public Double getSteuerbetrag() throws RemoteException;
 
 }
